change charts components to typescript
