+++ packages/client/package.json
@@ -73,6 +73,7 @@
     "@types/moment": "^2.13.0",
     "@types/react": "^16.7.18",
     "@types/react-dom": "^16.0.11",
+    "@types/react-router": "^4.4.3",
     "autoprefixer": "6.7.7",
     "awesome-typescript-loader": "^5.2.1",
     "babel-eslint": "^9.0.0",
